requirements.txt
langchain-core==0.3.21
langchain-openai==0.2.10
langgraph==0.2.53
pydantic==2.10.1
chainlit==1.3.2
python-dotenv==1.0.1